vfs: preserve main semantics and portability after compound rebase
Retain explicit-open grants, lock-stateid SETATTR, pNFS handle ownership, exclusive-create verifiers and directory stream reads across the merged execution paths. Separate capability bits and replace shared 128-bit range arithmetic with portable carry handling. Repair optimized-build fixtures and record the remaining validation failures.

+++ src/vfs/tests/vfs_clone_range_test.c
@@ -705,7 +705,7 @@ test_owned_native_gate(
     struct chimera_claim_actor      actor  = { .owner = { .proto = CHIMERA_CLAIM_PROTO_SMB2, .client_key = 456 } }
     ;
     struct chimera_vfs_module      *module       = dst->vfs_module;
-    unsigned int                    capabilities = module->capabilities;
+    uint64_t                        capabilities = module->capabilities;
 
     native_original       = module->dispatch;
     module->dispatch      = native_gate_dispatch;
